Move constant buffer info gathering logic too
